+++ edit-client/src/client/actions.rs
@@ -1,15 +1,16 @@
 mod styles;
 mod caret;
 mod modify;
+mod identify;
 
 pub use self::styles::*;
 pub use self::caret::*;
 pub use self::modify::*;
+pub use self::identify::*;
 use crate::walkers::*;
 use failure::Error;
 use oatie::doc::*;
 use oatie::OT;
-use std::collections::HashSet;
 
 fn is_boundary_char(c: char) -> bool {
     c.is_whitespace() || c == '-' || c == '_'
@@ -52,37 +53,5 @@ impl ActionContext {
 
     pub fn result(self) -> Op {
         self.op_result
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct CaretState {
-    pub block: String,
-    pub in_list: bool,
-    pub styles: HashSet<Style>,
-}
-
-// Return a "caret state".
-pub fn identify_block(ctx: ActionContext) -> Result<CaretState, Error> {
-    // Identify selection styles.
-    let styles = identify_styles(&ctx)?;
-
-    let mut walker = ctx.get_walker(Pos::Focus)?;
-    assert!(walker.back_block());
-    if let Some(DocGroup(ref attrs, _)) = walker.doc().head() {
-        let tag = attrs["tag"].clone();
-        let mut in_list = false;
-        if walker.parent() {
-            if let Some(DocGroup(ref attrs_2, _)) = walker.doc().head() {
-                in_list = attrs_2["tag"] == "bullet";
-            }
-        }
-        Ok(CaretState {
-            block: tag,
-            in_list,
-            styles,
-        })
-    } else {
-        bail!("Expected a DocGroup from back_block");
     }
 }
